Extracted default eval routines.
