+++ experiments/datakit/decontam/prepare_eval_corpus.py
@@ -54,6 +54,12 @@ from datasets import Image as DatasetsImage
 from datasets import load_dataset
 from huggingface_hub import hf_hub_download
 from marin.defaults import CORE_TASKS
+from marin.defaults.evals import (
+    KEY_GENERATION_TASKS,
+    KEY_MULTIPLE_CHOICE_TASKS,
+    OPEN_LM_LEADERBOARD_GEN,
+    OPEN_LM_LEADERBOARD_MCQ,
+)
 from rigging.filesystem import url_to_fs
 from rigging.log_setup import configure_logging
 
@@ -67,8 +73,6 @@ from experiments.evals.task_configs import (
     BIAS_SAFETY_TASKS,
     CODE_TASKS,
     EMOTIONAL_ETHICS_TASKS,
-    KEY_GENERATION_TASKS,
-    KEY_MULTIPLE_CHOICE_TASKS,
     KNOWLEDGE_TASKS,
     LANGUAGE_TASKS,
     MATH_TASKS,
@@ -76,8 +80,6 @@ from experiments.evals.task_configs import (
     MGSM_MULTILINGUAL_TASKS,
     MMLU_TASKS,
     MULTILINGUAL_LM_EVAL_LOGPROB_TASKS,
-    OPEN_LM_LEADERBOARD_GEN,
-    OPEN_LM_LEADERBOARD_MCQ,
     REASONING_TASKS,
     SPECIALIZED_TASKS,
     TRUTHFULNESS_TASKS,
