+++ experiments/evals/run_base_model_evals.py
@@ -9,9 +9,9 @@ MAP-NEO 7B, and Amber Base 7B models on CORE_TASKS (augmented with OLMo Eval Tas
 as dedicated MMLU 0-shot and 5-shot configurations.
 """
 
+from marin.defaults.evals import default_base_eval
 from marin.execution.executor import executor_main
 
-from experiments.evals.evals import default_base_eval
 from experiments.models import amber_base_7b, llama_3_1_8b, map_neo_7b, olmo_2_base_8b
 
 if __name__ == "__main__":
