+++ experiments/exp1994_32b_sft.py
@@ -10,9 +10,9 @@ import dataclasses
 
 from fray.cluster import ResourceConfig
 from marin.defaults import SimpleSFTConfig, default_sft
+from marin.defaults.evals import default_sft_eval
 from marin.execution.executor import executor_main
 
-from experiments.evals.evals import default_sft_eval
 from experiments.exp1880_sft_baseline import mixture_config
 from experiments.marin_models import marin_tokenizer
 from experiments.tootsie.exp1529_32b_mantis_cooldown import (
