+++ experiments/tootsie/exp1529_32b_bison_cooldown.py
@@ -18,10 +18,10 @@ from levanter.optim import AdamConfig
 from levanter.optim.clip_update_norm import ClipUpdateNormConfig
 from levanter.schedule import ScheduleStep
 from marin.defaults import default_train
+from marin.defaults.evals import default_base_eval
 from marin.execution import executor_main, output_path_of
 from marin.processing.tokenize.data_configs import lm_varying_mixture_data_config
 
-from experiments.evals.evals import default_base_eval
 from experiments.exp934_hq_vs_pt import pt_vs_hq_components
 from experiments.models import ModelConfig, download_model_step
 from experiments.pretraining_datasets import (
