+++ lib/marin/src/marin/defaults/__init__.py
@@ -5,8 +5,27 @@ from .config import SimpleDPOConfig, SimpleSFTConfig, SimpleTrainConfig
 from .evals import (
     ACTIVE_DATASETS,
     ALL_UNCHEATABLE_EVAL_DATASETS,
+    BASE_GENERATION_TASKS,
+    CORE_TASKS_PLUS_LEADERBOARD,
+    DEFAULT_LM_EVAL_MODEL_KWARGS,
+    DEFAULT_VLLM_ENGINE_KWARGS,
     EVAL_DEPENDENCY_GROUPS,
+    EVALCHEMY_DEPENDENCY_GROUPS,
+    KEY_GENERATION_TASKS,
+    KEY_MULTIPLE_CHOICE_TASKS,
+    MMLU_0_SHOT,
+    MMLU_5_SHOT,
+    MMLU_PRO_5_SHOT,
+    OPEN_LM_LEADERBOARD_GEN,
+    OPEN_LM_LEADERBOARD_MCQ,
+    default_base_eval,
     default_eval,
+    default_key_evals,
+    default_sft_eval,
+    evaluate_levanter_lm_evaluation_harness,
+    evaluate_lm_evaluation_harness,
+    extract_model_name_and_path,
+    uncheatable_eval,
     uncheatable_eval_tokenized,
 )
 from .step import CORE_TASKS, default_dpo, default_sft, default_train, default_validation_sets
