+++ lib/marin/src/marin/defaults/evals.py
@@ -42,6 +42,56 @@ ACTIVE_DATASETS = [
     "ao3_english",
 ]
 EVAL_DEPENDENCY_GROUPS = ["eval", "vllm", "tpu"]
+EVALCHEMY_DEPENDENCY_GROUPS = ["evalchemy", "vllm", "tpu"]
+
+
+DEFAULT_VLLM_ENGINE_KWARGS = {"max_model_len": 4096}
+DEFAULT_LM_EVAL_MODEL_KWARGS = {**DEFAULT_VLLM_ENGINE_KWARGS, "max_gen_toks": 4096}
+MMLU_0_SHOT = EvalTaskConfig("mmlu", 0, task_alias="mmlu_0shot")
+MMLU_5_SHOT = EvalTaskConfig("mmlu", 5, task_alias="mmlu_5shot")
+MMLU_PRO_5_SHOT = EvalTaskConfig("leaderboard_mmlu_pro", 5, task_alias="mmlu_5shot")
+OPEN_LM_LEADERBOARD_MCQ = (
+    EvalTaskConfig("leaderboard_bbh", 3, task_alias="lb_bbh_3shot"),
+    EvalTaskConfig("leaderboard_mmlu_pro", 5, task_alias="lb_mmlu_pro_5shot"),
+    EvalTaskConfig("leaderboard_gpqa", 0, task_alias="lb_gpqa_0shot"),
+    EvalTaskConfig("leaderboard_musr", 0, task_alias="lb_musr_0shot"),
+)
+OPEN_LM_LEADERBOARD_GEN = (
+    EvalTaskConfig("leaderboard_ifeval", 0, task_alias="lb_ifeval_0shot"),
+    EvalTaskConfig("leaderboard_math_hard", 4, task_alias="lb_math_4shot"),
+)
+CORE_TASKS_PLUS_LEADERBOARD = (
+    EvalTaskConfig(
+        "leaderboard_bbh",
+        3,
+        task_alias="bbh_3shot",
+    ),
+    EvalTaskConfig(
+        "leaderboard_gpqa",
+        0,
+        task_alias="gpqa_0shot",
+    ),
+    *CORE_TASKS,
+)
+BASE_GENERATION_TASKS = (
+    EvalTaskConfig(name="bbh_cot_fewshot", num_fewshot=3),
+    EvalTaskConfig(name="gsm8k_cot", num_fewshot=8),
+    EvalTaskConfig(name="nq_open", num_fewshot=0, task_alias="nq_open"),
+    EvalTaskConfig(name="triviaqa", num_fewshot=0, task_alias="triviaqa"),
+)
+KEY_GENERATION_TASKS = (
+    EvalTaskConfig(name="ifeval", num_fewshot=0),
+    EvalTaskConfig(name="gsm8k_cot", num_fewshot=8),
+    EvalTaskConfig(name="drop", num_fewshot=0),
+    EvalTaskConfig(name="humaneval", num_fewshot=10),
+    EvalTaskConfig(name="bbh_cot_fewshot", num_fewshot=3, task_alias="bbh"),
+    EvalTaskConfig(name="minerva_math", num_fewshot=4, task_alias="math_4shot"),
+)
+KEY_MULTIPLE_CHOICE_TASKS = (
+    EvalTaskConfig("mmlu", 0, task_alias="mmlu_0shot"),
+    EvalTaskConfig("mmlu", 5, task_alias="mmlu_5shot"),
+    EvalTaskConfig(name="truthfulqa_mc2", num_fewshot=6, task_alias="truthqa"),
+)
 
 
 uncheatable_eval = make_uncheatable_eval_step()
@@ -136,6 +186,55 @@ def evaluate_levanter_lm_evaluation_harness(
     )
 
 
+def evaluate_lm_evaluation_harness(
+    model_name: str,
+    model_path: str,
+    evals: list[EvalTaskConfig],
+    max_eval_instances: int | None = None,
+    engine_kwargs: dict | None = None,
+    resource_config: ResourceConfig | None = None,
+    apply_chat_template: bool = False,
+    wandb_tags: list[str] | None = None,
+    discover_latest_checkpoint: bool = True,
+    env_vars: dict[str, str] | None = None,
+) -> ExecutorStep:
+    """
+    Create an ExecutorStep to evaluate the model using LM Evaluation Harness.
+
+    Args:
+        model_name (str): Name of the model.
+        model_path (str): Path to the model.
+        evals (list[EvalTaskConfig]): List of evaluations to run with LM Evaluation Harness.
+        env_vars (dict[str, str] | None): Extra env vars to set on the child iris worker.
+            Needed for vLLM-on-TPU bring-up (e.g. ``VLLM_ENABLE_V1_MULTIPROCESSING=0``)
+            and code-eval-dependent tasks like humaneval (``HF_ALLOW_CODE_EVAL=1``).
+            The coordinator's own ``os.environ`` does NOT propagate to iris-spawned
+            children — these vars must be threaded through ``remote()``.
+    """
+    return ExecutorStep(
+        name=f"evaluation/lm_evaluation_harness/{model_name}",
+        fn=remote(
+            evaluate,
+            resources=resource_config,
+            pip_dependency_groups=EVAL_DEPENDENCY_GROUPS,
+            env_vars=env_vars,
+        ),
+        config=EvaluationConfig(
+            evaluator="lm_evaluation_harness",
+            model_name=model_name,
+            model_path=model_path,
+            evaluation_path=this_output_path(),
+            evals=evals,
+            max_eval_instances=max_eval_instances,
+            discover_latest_checkpoint=discover_latest_checkpoint,
+            engine_kwargs=engine_kwargs,
+            resource_config=resource_config,
+            apply_chat_template=apply_chat_template,
+            wandb_tags=wandb_tags,
+        ),
+    )
+
+
 def default_eval(
     step: ExecutorStep | InputName | str,
     resource_config: ResourceConfig = ResourceConfig.with_tpu("v4-8"),
@@ -173,3 +272,182 @@ def default_eval(
         apply_chat_template=apply_chat_template,
         discover_latest_checkpoint=discover_latest_checkpoint,
     )
+
+
+def default_base_eval(
+    step: ExecutorStep | InputName | str,
+    resource_config: ResourceConfig = ResourceConfig.with_tpu("v6e-8"),
+    max_eval_instances: int | None = None,
+    engine_kwargs: dict | None = DEFAULT_LM_EVAL_MODEL_KWARGS,
+    run_generation_evals: bool = True,
+    discover_latest_checkpoint: bool = True,
+):
+    # Add GPQA to CORE_TASKS
+    # Set up evaluations for core tasks (including GPQA)
+    eval_jobs = []
+    core_grouped = default_eval(
+        step=step,
+        resource_config=resource_config,
+        evals=CORE_TASKS_PLUS_LEADERBOARD,
+        discover_latest_checkpoint=discover_latest_checkpoint,
+    )
+    eval_jobs.append(core_grouped)
+
+    # Run tasks where we report Macro_Avg separately to make sure the macro avg gets computed correctly.
+    mmlu_0shot = default_eval(
+        step=step,
+        resource_config=resource_config,
+        evals=(MMLU_0_SHOT,),
+        discover_latest_checkpoint=discover_latest_checkpoint,
+    )
+    eval_jobs.append(mmlu_0shot)
+
+    mmlu_5shot = default_eval(
+        step=step,
+        resource_config=resource_config,
+        evals=(MMLU_5_SHOT,),
+        discover_latest_checkpoint=discover_latest_checkpoint,
+    )
+    eval_jobs.append(mmlu_5shot)
+
+    mmlu_pro_5shot = default_eval(
+        step=step,
+        resource_config=resource_config,
+        evals=(MMLU_PRO_5_SHOT,),
+        discover_latest_checkpoint=discover_latest_checkpoint,
+    )
+    eval_jobs.append(mmlu_pro_5shot)
+
+    name, model_step_path = extract_model_name_and_path(step)
+    if run_generation_evals:
+        generation = evaluate_lm_evaluation_harness(
+            name,
+            model_step_path,
+            BASE_GENERATION_TASKS,
+            max_eval_instances=max_eval_instances,
+            engine_kwargs=engine_kwargs,
+            resource_config=resource_config,
+            discover_latest_checkpoint=discover_latest_checkpoint,
+        )
+
+        eval_jobs.append(generation)
+    return eval_jobs
+
+
+def default_sft_eval(
+    step: ExecutorStep | InputName | str,
+    resource_config: ResourceConfig = ResourceConfig.with_tpu("v6e-8"),
+    max_eval_instances: int | None = None,
+    engine_kwargs: dict | None = DEFAULT_LM_EVAL_MODEL_KWARGS,
+    run_generation_evals: bool = True,
+    apply_chat_template: bool = True,
+    use_levanter_inference: bool = False,
+):
+    # Set up evaluations for core tasks (including GPQA)
+    eval_jobs = []
+    leaderboard_grouped = default_eval(
+        step=step,
+        resource_config=resource_config,
+        evals=OPEN_LM_LEADERBOARD_MCQ,
+        apply_chat_template=apply_chat_template,
+    )
+    eval_jobs.append(leaderboard_grouped)
+
+    # Run tasks where we report Macro_Avg separately to make sure the macro avg gets computed correctly.
+
+    mmlu_5shot = default_eval(
+        step=step, resource_config=resource_config, evals=(MMLU_5_SHOT,), apply_chat_template=apply_chat_template
+    )
+    eval_jobs.append(mmlu_5shot)
+
+    mmlu_pro_5shot = default_eval(
+        step=step, resource_config=resource_config, evals=(MMLU_PRO_5_SHOT,), apply_chat_template=apply_chat_template
+    )
+    eval_jobs.append(mmlu_pro_5shot)
+
+    name, model_step_path = extract_model_name_and_path(step)
+    if run_generation_evals:
+        if use_levanter_inference:
+            leaderboard_generation = evaluate_levanter_lm_evaluation_harness(
+                name,
+                model_step_path,
+                KEY_GENERATION_TASKS,
+                resource_config,
+                max_eval_instances=max_eval_instances,
+                apply_chat_template=apply_chat_template,
+            )
+            eval_jobs.append(leaderboard_generation)
+
+            olmo_generation = evaluate_levanter_lm_evaluation_harness(
+                name,
+                model_step_path,
+                OPEN_LM_LEADERBOARD_GEN,
+                resource_config,
+                max_eval_instances=max_eval_instances,
+                apply_chat_template=apply_chat_template,
+            )
+            eval_jobs.append(olmo_generation)
+        else:
+            leaderboard_generation = evaluate_lm_evaluation_harness(
+                name,
+                model_step_path,
+                KEY_GENERATION_TASKS,
+                max_eval_instances=max_eval_instances,
+                engine_kwargs=engine_kwargs,
+                resource_config=resource_config,
+                apply_chat_template=apply_chat_template,
+            )
+
+            eval_jobs.append(leaderboard_generation)
+
+            olmo_generation = evaluate_lm_evaluation_harness(
+                name,
+                model_step_path,
+                OPEN_LM_LEADERBOARD_GEN,
+                max_eval_instances=max_eval_instances,
+                engine_kwargs=engine_kwargs,
+                resource_config=resource_config,
+                apply_chat_template=apply_chat_template,
+            )
+            eval_jobs.append(olmo_generation)
+    return eval_jobs
+
+
+def default_key_evals(
+    step: ExecutorStep | InputName | str,
+    resource_config: ResourceConfig,
+    model_name: str | None = None,
+    max_eval_instances: int | None = None,
+    engine_kwargs: dict | None = DEFAULT_LM_EVAL_MODEL_KWARGS,
+) -> list[ExecutorStep]:
+    """
+    Create a list of ExecutorSteps to evaluate the model using LM Evaluation Harness on a step.
+    """
+    name, model_step_path = extract_model_name_and_path(step)
+
+    if model_name is None:
+        model_name = name
+
+    stop_token_ids = []
+    if "llama3" in model_name:
+        stop_token_ids.append(128009)
+    elif "olmo" in model_name:
+        stop_token_ids.append(100257)
+
+    return [
+        evaluate_lm_evaluation_harness(
+            model_name,
+            model_step_path,
+            KEY_GENERATION_TASKS,
+            max_eval_instances=max_eval_instances,
+            engine_kwargs=engine_kwargs,
+            resource_config=resource_config,
+        ),
+        evaluate_levanter_lm_evaluation_harness(
+            model_name,
+            model_step_path,
+            KEY_MULTIPLE_CHOICE_TASKS,
+            resource_config,
+            max_eval_instances=max_eval_instances,
+        ),
+    ]
